+++ lib/rubyzen/declarations/method_declaration.rb
@@ -21,6 +21,8 @@ module Rubyzen
       include Rubyzen::Providers::VisibilityProvider
       include Rubyzen::Providers::RescuesProvider
       include Rubyzen::Providers::RaisesProvider
+      include Rubyzen::Providers::ReturnsProvider
+      include Rubyzen::Providers::AssignmentsProvider
 
       # @return [RuboCop::AST::Node]
       attr_reader :node