+++ lib/rubyzen/providers/arguments_provider.rb
@@ -1,0 +1,15 @@
+module Rubyzen
+  module Providers
+    # Provides the arguments passed at a call site (or macro), as expressions.
+    module ArgumentsProvider
+      # @return [Rubyzen::Collections::ArgumentsCollection]
+      def arguments
+        Collections::ArgumentsCollection.new(
+          node.arguments.map do |argument_node|
+            Declarations::ExpressionDeclaration.new(argument_node, self)
+          end
+        )
+      end
+    end
+  end
+end